+++ Editor/Utility/ScriptableObjectUtility.cs
@@ -1,7 +1,6 @@
 using UnityEngine;
 using UnityEditor;
 using System.IO;
-using System.Linq;
 using System;
 
 namespace Cinemachine.Editor
@@ -10,6 +9,8 @@ namespace Cinemachine.Editor
     {
         public static string kPackageRoot = "Packages/com.unity.cinemachine";
 
+        /// <summary>Get the Cinemachine package install path.  Works whether CM is
+        /// a packman package or an ordinary asset.</summary>
         public static string CinemachineInstallPath
         {
             get 
@@ -36,11 +37,13 @@ namespace Cinemachine.Editor
             }
         }
 
+        /// <summary>Create a scriptable object asset</summary>
         public static T CreateAt<T>(string assetPath) where T : ScriptableObject
         {
             return CreateAt(typeof(T), assetPath) as T;
         }
 
+        /// <summary>Create a scriptable object asset</summary>
         public static ScriptableObject CreateAt(Type assetType, string assetPath)
         {
             ScriptableObject asset = ScriptableObject.CreateInstance(assetType);
@@ -52,46 +55,6 @@ namespace Cinemachine.Editor
             AssetDatabase.CreateAsset(asset, assetPath);
             return asset;        
         }
-
-
-#if false // unused
-        public static bool AddDefineForAllBuildTargets(string k_Define)
-        {
-            bool added = false;
-            var targets = Enum.GetValues(typeof(BuildTargetGroup))
-                .Cast<BuildTargetGroup>()
-                .Where(x => x != BuildTargetGroup.Unknown)
-                .Where(x => !BuildTargetIsObsolete(x));
-
-            foreach (var target in targets)
-            {
-                var defines = PlayerSettings.GetScriptingDefineSymbolsForGroup(target).Trim();
-
-                var list = defines.Split(';', ' ')
-                    .Where(x => !string.IsNullOrEmpty(x))
-                    .ToList();
-
-                if (!list.Contains(k_Define))
-                {
-                    list.Add(k_Define);
-                    defines = list.Aggregate((a, b) => a + ";" + b);
-
-                    PlayerSettings.SetScriptingDefineSymbolsForGroup(target, defines);
-                    added = true;
-                }
-            }
-            return added;
-        }
-
-        static bool BuildTargetIsObsolete(BuildTargetGroup group)
-        {
-            var attrs = typeof(BuildTargetGroup)
-                .GetField(group.ToString())
-                .GetCustomAttributes(typeof(ObsoleteAttribute), false);
-
-            return attrs != null && attrs.Length > 0;
-        }
-#endif
 
         public static void Create<T>(bool prependFolderName = false, bool trimName = true) where T : ScriptableObject
         {